Separate react-native related logic from main code
