+++ lib/cache.js
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import store from './store';
+import getStore from './store';
 
 export default class Cache {
 
@@ -21,7 +21,7 @@ export default class Cache {
     this._maxRetryCount = 1;
     this.prefix = prefix;
     this.map = {};
-    this.store = store;
+    this.store = getStore();
   }
 
   _applyNamespaceOnKey(key) {
