+++ lib/container.js
@@ -16,7 +16,6 @@
 /* eslint camelcase: 0 */
 const request = require('superagent');
 const _ = require('lodash');
-const store = require('./store');
 const ee = require('event-emitter');
 
 import Asset from './asset';
@@ -30,14 +29,10 @@ import Database from './database';
 import Pubsub from './pubsub';
 import {RelationAction} from './relation';
 import Geolocation from './geolocation';
+import getStore from './store';
 import {Sequence} from './type';
 import {ErrorCodes, SkygearError} from './error';
 import {EventHandle} from './util';
-
-var React;
-if (typeof window !== 'undefined') {
-  React = require('react-native');
-}
 
 export const USER_CHANGED = 'userChanged';
 
@@ -104,7 +99,7 @@ export default class Container {
   }
 
   clearCache() {
-    return store.clearPurgeableItems();
+    return this.store.clearPurgeableItems();
   }
 
   onUserChanged(listener) {
@@ -414,6 +409,12 @@ export default class Container {
     });
   }
 
+  inferDeviceType() {
+    // To be implmented by subclass
+    // TODO: probably web / node, handle it later
+    throw new Error('Failed to infer type, please supply a value');
+  }
+
   /**
    * You can register your device for receiving push notifications.
    *
@@ -426,18 +427,8 @@ export default class Container {
     if (!token) {
       throw new Error('Token cannot be empty');
     }
-
     if (!type) {
-      if (React) {
-        if (React.Platform.OS === 'ios') {
-          type = 'ios';
-        } else {
-          type = 'android';
-        }
-      } else {
-        // TODO: probably web / node, handle it later
-        throw new Error('Failed to infer type, please supply a value');
-      }
+      type = this.inferDeviceType();
     }
 
     let deviceID;
@@ -672,7 +663,7 @@ export default class Container {
 
   _getUser() {
     let self = this;
-    return store.getItem('skygear-user').then(function (userJSON) {
+    return this.store.getItem('skygear-user').then(function (userJSON) {
       let attrs = JSON.parse(userJSON);
       self._user = self.User.fromJSON(attrs);
     }, function (err) {
@@ -693,8 +684,8 @@ export default class Container {
       value = null;
     }
 
-    const setItem = value === null ? store.removeItem('skygear-user')
-        : store.setItem('skygear-user', value);
+    const setItem = value === null ? this.store.removeItem('skygear-user')
+        : this.store.setItem('skygear-user', value);
     return setItem.then(function () {
       container.ee.emit(USER_CHANGED, container._user);
       return value;
@@ -710,7 +701,7 @@ export default class Container {
 
   _getAccessToken() {
     let self = this;
-    return store.getItem('skygear-accesstoken').then(function (token) {
+    return this.store.getItem('skygear-accesstoken').then(function (token) {
       self._accessToken = token;
       return token;
     }, function (err) {
@@ -722,8 +713,9 @@ export default class Container {
 
   _setAccessToken(value) {
     this._accessToken = value;
-    const setItem = value === null ? store.removeItem('skygear-accesstoken')
-        : store.setItem('skygear-accesstoken', value);
+    const setItem = value === null
+        ? this.store.removeItem('skygear-accesstoken')
+        : this.store.setItem('skygear-accesstoken', value);
     return setItem.then(function () {
       return value;
     }, function (err) {
@@ -738,7 +730,7 @@ export default class Container {
 
   _getDeviceID() {
     let self = this;
-    return store.getItem('skygear-deviceid').then(function (deviceID) {
+    return this.store.getItem('skygear-deviceid').then(function (deviceID) {
       self._deviceID = deviceID;
       return deviceID;
     }, function (err) {
@@ -751,8 +743,8 @@ export default class Container {
   _setDeviceID(value) {
     let self = this;
     this._deviceID = value;
-    const setItem = value === null ? store.removeItem('skygear-deviceid')
-        : store.setItem('skygear-deviceid', value);
+    const setItem = value === null ? this.store.removeItem('skygear-deviceid')
+        : this.store.setItem('skygear-deviceid', value);
     return setItem.then(function () {
       return value;
     }, function (err) {
@@ -773,6 +765,13 @@ export default class Container {
     if (newEndPoint) {
       this.url = newEndPoint;
     }
+  }
+
+  get store() {
+    if (!this._store) {
+      this._store = getStore();
+    }
+    return this._store;
   }
 
   get publicDB() {
