+++ lib/react-native/index.js
@@ -1,0 +1,17 @@
+import reactNative from 'react-native';
+import Container from '../container';
+import {setStore} from '../store';
+import reactNativeStore from './store';
+
+class ReactNativeContainer extends Container {
+  inferDeviceType() {
+    if (reactNative.Platform.OS === 'ios') {
+      return 'ios';
+    }
+    return 'android';
+  }
+}
+
+setStore(reactNativeStore);
+
+export default new ReactNativeContainer();
