+++ lib/store.js
@@ -166,95 +166,6 @@ class SyncStorageDriver {
 
 }
 
-class ReactNativeAsyncStorageDriver {
-
-  constructor(rnImpl) {
-    this._rnImpl = rnImpl;
-  }
-
-  clear(callback) {
-    return this._rnImpl.clear(callback);
-  }
-
-  getItem(key, callback) {
-    return this._rnImpl.getItem(key, callback);
-  }
-
-  setItem(key, value, callback) {
-    return this._rnImpl.setItem(key, value, callback);
-  }
-
-  removeItem(key, callback) {
-    return this._rnImpl.removeItem(key, callback);
-  }
-
-  multiGet(keys, callback) {
-    return this._rnImpl.multiGet(keys).then(function (rnKeyValuePairs) {
-      const output = [];
-      for (let i = 0; i < rnKeyValuePairs.length; ++i) {
-        const rnPair = rnKeyValuePairs[i];
-        const key = rnPair[0];
-        const value = rnPair[1];
-        output.push({
-          key: key,
-          value: value
-        });
-      }
-      if (callback) {
-        callback(null, output);
-      }
-      return output;
-    }, function (errors) {
-      if (callback) {
-        callback(errors);
-      }
-      return Promise.reject(errors);
-    });
-  }
-
-  multiSet(keyValuePairs, callback) {
-    const rnKeyValuePairs = [];
-    for (let i = 0; i < keyValuePairs.length; ++i) {
-      const pair = keyValuePairs[i];
-      const key = pair.key;
-      const value = pair.value;
-      rnKeyValuePairs.push([key, value]);
-    }
-    return this._rnImpl.multiSet(rnKeyValuePairs, callback);
-  }
-
-  multiRemove(keys, callback) {
-    return this._rnImpl.multiRemove(keys, callback);
-  }
-
-  key(n, callback) {
-    return this._rnImpl.getAllKeys().then(function (allKeys) {
-      let result = null;
-      if (n >= 0 && n < allKeys.length) {
-        result = allKeys[n];
-      }
-      if (callback) {
-        callback(null, result);
-      }
-      return result;
-    });
-  }
-
-  keys(callback) {
-    return this._rnImpl.getAllKeys(callback);
-  }
-
-  length(callback) {
-    return this._rnImpl.getAllKeys().then(function (allKeys) {
-      if (callback) {
-        callback(null, allKeys.length);
-      }
-      return allKeys.length;
-    });
-  }
-
-}
-
 class Store {
   constructor(driver, keyWhiteList) {
     this._driver = driver;
@@ -469,25 +380,29 @@ class Store {
 
 }
 
-/* global window: false */
-if (typeof window !== 'undefined') {
-  // env: browser-like
-  var rn = require('react-native');
-  if (rn && rn.AsyncStorage) {
-    // env: ReactNative
-    store = new Store(new ReactNativeAsyncStorageDriver(rn.AsyncStorage));
-  } else if (isLocalStorageValid()) {
-    // env: Modern browsers
-    store = new Store(new SyncStorageDriver(window.localStorage));
-  } else {
-    // env: Legacy browsers
-    var cookieImpl = new CookieStorage();
-    store = new Store(new SyncStorageDriver(cookieImpl, cookieKeyWhiteList));
-  }
-} else {
-  // env: node
-  var memoryImpl = require('localstorage-memory');
-  store = new Store(new SyncStorageDriver(memoryImpl));
-}
+export const setStore = (_store) => {
+  store = _store;
+};
 
-export default store;
+export default ()=> {
+  if (store) {
+    return store;
+  }
+  /* global window: false */
+  if (typeof window !== 'undefined') {
+    // env: browser-like
+    if (isLocalStorageValid()) {
+      // env: Modern browsers
+      store = new Store(new SyncStorageDriver(window.localStorage));
+    } else {
+      // env: Legacy browsers
+      var cookieImpl = new CookieStorage();
+      store = new Store(new SyncStorageDriver(cookieImpl, cookieKeyWhiteList));
+    }
+  } else {
+    // env: node
+    var memoryImpl = require('localstorage-memory');
+    store = new Store(new SyncStorageDriver(memoryImpl));
+  }
+  return store;
+};
